Added tableName for Tag & Post models, added controller and router for post table, added getAll post API, updated app.js, re-adjust seeder file for post user_id (remote change)
