+++ db/models/Tag.js
@@ -1,7 +1,5 @@
-'use strict';
-const {
-  Model
-} = require('sequelize');
+"use strict";
+const { Model } = require("sequelize");
 module.exports = (sequelize, DataTypes) => {
   class Tag extends Model {
     /**
@@ -16,20 +14,25 @@ module.exports = (sequelize, DataTypes) => {
         through: "TagList",
         foreignKey: "tag_name",
         timestamps: false,
-      })
+      });
     }
   }
-  Tag.init({
-    name: {
-      type: DataTypes.STRING,
-      allowNull: false,
-      primaryKey: true,
+  Tag.init(
+    {
+      name: {
+        type: DataTypes.STRING,
+        allowNull: false,
+        primaryKey: true,
+      },
     },
-  }, {
-    sequelize,
-    modelName: 'Tag',
-    underscored: true,
-    timestamps: false,
-  });
+    {
+      sequelize,
+      modelName: "Tag",
+      underscored: true,
+      tableName: "Tags",
+
+      timestamps: false,
+    }
+  );
   return Tag;
 };
